organize methods
