+++ ocpmodels/trainers/base_trainer.py
@@ -30,10 +30,8 @@ from ocpmodels.common.registry import registry
 from ocpmodels.common.typing import assert_is_instance as aii
 from ocpmodels.common.typing import none_throws
 from ocpmodels.common.utils import (
-    cg_change_mat,
     get_commit_hash,
     get_loss_module,
-    irreps_sum,
     load_state_dict,
     save_checkpoint,
     update_config,
@@ -691,283 +689,6 @@ class BaseTrainer(ABC):
                     results_file="predictions",
                     disable_tqdm=disable_eval_tqdm,
                 )
-
-    def train(self, disable_eval_tqdm: bool = False) -> None:
-        ensure_fitted(self._unwrapped_model, warn=True)
-
-        eval_every = self.config["optim"].get(
-            "eval_every", len(self.train_loader)
-        )
-        checkpoint_every = self.config["optim"].get(
-            "checkpoint_every", eval_every
-        )
-        primary_metric = self.evaluation_metrics.get(
-            "primary_metric", self.evaluator.task_primary_metric[self.name]
-        )
-        if (
-            not hasattr(self, "primary_metric")
-            or self.primary_metric != primary_metric
-        ):
-            self.best_val_metric = 1e9 if "mae" in primary_metric else -1.0
-        else:
-            primary_metric = self.primary_metric
-        self.metrics = {}
-
-        # Calculate start_epoch from step instead of loading the epoch number
-        # to prevent inconsistencies due to different batch size in checkpoint.
-        start_epoch = self.step // len(self.train_loader)
-
-        for epoch_int in range(
-            start_epoch, self.config["optim"]["max_epochs"]
-        ):
-            self.train_sampler.set_epoch(epoch_int)
-            skip_steps = self.step % len(self.train_loader)
-            train_loader_iter = iter(self.train_loader)
-
-            for i in range(skip_steps, len(self.train_loader)):
-                self.epoch = epoch_int + (i + 1) / len(self.train_loader)
-                self.step = epoch_int * len(self.train_loader) + i + 1
-                self.model.train()
-
-                # Get a batch.
-                batch = next(train_loader_iter)
-
-                # Forward, loss, backward.
-                with torch.cuda.amp.autocast(enabled=self.scaler is not None):
-                    out = self._forward(batch)
-                    loss = self._compute_loss(out, batch)
-                loss = self.scaler.scale(loss) if self.scaler else loss
-                self._backward(loss)
-                scale = self.scaler.get_scale() if self.scaler else 1.0
-
-                # Compute metrics.
-                self.metrics = self._compute_metrics(
-                    out,
-                    batch,
-                    self.evaluator,
-                    self.metrics,
-                )
-                self.metrics = self.evaluator.update(
-                    "loss", loss.item() / scale, self.metrics
-                )
-
-                # Log metrics.
-                log_dict = {k: self.metrics[k]["metric"] for k in self.metrics}
-                log_dict.update(
-                    {
-                        "lr": self.scheduler.get_lr(),
-                        "epoch": self.epoch,
-                        "step": self.step,
-                    }
-                )
-                if (
-                    self.step % self.config["cmd"]["print_every"] == 0
-                    and distutils.is_master()
-                ):
-                    log_str = [
-                        "{}: {:.2e}".format(k, v) for k, v in log_dict.items()
-                    ]
-                    logging.info(", ".join(log_str))
-                    self.metrics = {}
-
-                if self.logger is not None:
-                    self.logger.log(
-                        log_dict,
-                        step=self.step,
-                        split="train",
-                    )
-
-                if (
-                    checkpoint_every != -1
-                    and self.step % checkpoint_every == 0
-                ):
-                    self.save(
-                        checkpoint_file="checkpoint.pt", training_state=True
-                    )
-
-                # Evaluate on val set every `eval_every` iterations.
-                if self.step % eval_every == 0:
-                    if self.val_loader is not None:
-                        val_metrics = self.validate(
-                            split="val",
-                            disable_tqdm=disable_eval_tqdm,
-                        )
-                        self.update_best(
-                            primary_metric,
-                            val_metrics,
-                            disable_eval_tqdm=disable_eval_tqdm,
-                        )
-
-                    if self.config["task"].get("eval_relaxations", False):
-                        if "relax_dataset" not in self.config["task"]:
-                            logging.warning(
-                                "Cannot evaluate relaxations, relax_dataset not specified"
-                            )
-                        else:
-                            self.run_relaxations()
-
-                if self.scheduler.scheduler_type == "ReduceLROnPlateau":
-                    if self.step % eval_every == 0:
-                        self.scheduler.step(
-                            metrics=val_metrics[primary_metric]["metric"],
-                        )
-                else:
-                    self.scheduler.step()
-
-            torch.cuda.empty_cache()
-
-            if checkpoint_every == -1:
-                self.save(checkpoint_file="checkpoint.pt", training_state=True)
-
-        self.train_dataset.close_db()
-        if self.config.get("val_dataset", False):
-            self.val_dataset.close_db()
-        if self.config.get("test_dataset", False):
-            self.test_dataset.close_db()
-
-    def _forward(self, batch):
-        out = self.model(batch.to(self.device))
-
-        ### TOOD: Move into BaseModel in OCP 2.0
-        outputs = {}
-        batch_size = batch.natoms.numel()
-        for target_key in self.config["outputs"]:
-            ### Target property is a direct output of the model
-            if target_key in out:
-                pred = out[target_key]
-            ## Target property is a derived output of the model. Construct the
-            ## parent property
-            else:
-                _max_rank = 0
-                for subtarget_key in self.config["outputs"][target_key][
-                    "decomposition"
-                ]:
-                    _max_rank = max(
-                        _max_rank,
-                        self.output_targets[subtarget_key]["irrep_dim"],
-                    )
-
-                pred_irreps = torch.zeros(
-                    (batch_size, irreps_sum(_max_rank)), device=self.device
-                )
-
-                for subtarget_key in self.config["outputs"][target_key][
-                    "decomposition"
-                ]:
-                    irreps = self.output_targets[subtarget_key]["irrep_dim"]
-                    _pred = out[subtarget_key]
-
-                    ## Fill in the corresponding irreps prediction
-                    pred_irreps[
-                        :,
-                        max(0, irreps_sum(irreps - 1)) : irreps_sum(irreps),
-                    ] = _pred
-
-                pred = torch.einsum(
-                    "ba, cb->ca",
-                    cg_change_mat(_max_rank, self.device),
-                    pred_irreps,
-                )
-
-            ### not all models are consistent with the output shape
-            if len(pred.shape) > 1:
-                pred = pred.squeeze(1)
-
-            outputs[target_key] = pred
-
-        return outputs
-
-    def _compute_loss(self, out, batch):
-        natoms = batch.natoms.to(self.device)
-        batch_size = natoms.numel()
-        natoms = torch.repeat_interleave(natoms, natoms)
-
-        fixed = batch.fixed.to(self.device)
-        mask = fixed == 0
-
-        loss = []
-
-        for loss_fn in self.loss_fns:
-            target_name, loss_info = loss_fn
-
-            target = batch[target_name].to(self.device)
-            pred = out[target_name]
-
-            if self.output_targets[target_name].get(
-                "level", "system"
-            ) == "atom" and self.output_targets[target_name].get(
-                "train_on_free_atoms", True
-            ):
-                target = target[mask]
-                pred = pred[mask]
-                natoms = natoms[mask]
-
-            if self.normalizers.get(target_name, False):
-                target = self.normalizers[target_name].norm(target)
-
-            mult = loss_info["coefficient"]
-
-            loss.append(
-                mult
-                * loss_info["fn"](
-                    pred,
-                    target,
-                    natoms=natoms,
-                    batch_size=batch_size,
-                )
-            )
-
-        # Sanity check to make sure the compute graph is correct.
-        for lc in loss:
-            assert hasattr(lc, "grad_fn")
-
-        loss = sum(loss)
-        return loss
-
-    def _compute_metrics(self, out, batch, evaluator, metrics={}):
-        natoms = batch.natoms.to(self.device)
-
-        ### Retrieve free atoms
-        fixed = batch.fixed.to(self.device)
-        mask = fixed == 0
-
-        s_idx = 0
-        natoms_free = []
-        for _natoms in natoms:
-            natoms_free.append(torch.sum(mask[s_idx : s_idx + _natoms]).item())
-            s_idx += _natoms
-        natoms = torch.LongTensor(natoms_free).to(self.device)
-
-        targets = {}
-        for target_name in self.output_targets:
-            target = batch[target_name].to(self.device)
-            # Add parent target to targets
-            if "parent" in self.output_targets[target_name]:
-                parent_target_name = self.output_targets[target_name]["parent"]
-
-                if parent_target_name not in targets:
-                    parent_target = batch[parent_target_name].to(self.device)
-                    targets[parent_target_name] = parent_target
-
-            if self.output_targets[target_name].get(
-                "level", "system"
-            ) == "atom" and self.output_targets[target_name].get(
-                "eval_on_free_atoms", True
-            ):
-                target = target[mask]
-                out[target_name] = out[target_name][mask]
-
-            targets[target_name] = target
-            if self.normalizers.get(target_name, False):
-                out[target_name] = self.normalizers[target_name].denorm(
-                    out[target_name]
-                )
-
-        targets["natoms"] = natoms
-        out["natoms"] = natoms
-
-        metrics = evaluator.eval(out, targets, prev_metrics=metrics)
-        return metrics
 
     @torch.no_grad()
     def validate(self, split: str = "val", disable_tqdm: bool = False):
